+++ src/Stack/Build.hs
@@ -72,7 +72,7 @@ build setLocalFiles mbuildLk bopts = do
            $ Set.unions
            $ map lpFiles locals
 
-    (installedMap, locallyRegistered) <-
+    (installedMap, globallyRegistered, locallyRegistered) <-
         getInstalled menv
                      GetInstalledOpts
                          { getInstalledProfiling = profiling
@@ -98,7 +98,11 @@ build setLocalFiles mbuildLk bopts = do
 
     if boptsDryrun bopts
         then printPlan plan
-        else executePlan menv bopts baseConfigOpts locals sourceMap installedMap plan
+        else executePlan menv bopts baseConfigOpts locals
+                         globallyRegistered
+                         sourceMap
+                         installedMap
+                         plan
   where
     profiling = boptsLibProfile bopts || boptsExeProfile bopts
 
@@ -142,7 +146,12 @@ withLoadPackage menv inner = do
     withCabalLoader menv $ \cabalLoader ->
         inner $ \name version flags -> do
             bs <- cabalLoader $ PackageIdentifier name version -- TODO automatically update index the first time this fails
-            readPackageBS (depPackageConfig econfig flags) bs
+
+            -- Intentionally ignore warnings, as it's not really
+            -- appropriate to print a bunch of warnings out while
+            -- resolving the package index.
+            (_warnings,pkg) <- readPackageBS (depPackageConfig econfig flags) bs
+            return pkg
   where
     -- | Package config to be used for dependencies
     depPackageConfig :: EnvConfig -> Map FlagName Bool -> PackageConfig