+++ src/Stack/Types/Compiler.hs
@@ -5,9 +5,8 @@
 module Stack.Types.Compiler where
 
 import           Control.DeepSeq
-import           Control.DeepSeq.Generics (genericRnf)
 import           Data.Aeson
-import           Data.Binary (Binary)
+import           Data.Binary.VersionTagged (Binary, HasStructuralInfo)
 import           Data.Monoid ((<>))
 import qualified Data.Text as T
 import           GHC.Generics (Generic)
@@ -34,8 +33,8 @@ data CompilerVersion
         {-# UNPACK #-} !Version -- GHC version
     deriving (Generic, Show, Eq, Ord)
 instance Binary CompilerVersion
-instance NFData CompilerVersion where
-    rnf = genericRnf
+instance HasStructuralInfo CompilerVersion
+instance NFData CompilerVersion
 instance ToJSON CompilerVersion where
     toJSON = toJSON . compilerVersionName
 instance FromJSON CompilerVersion where