+++ src/Stack/Types/FlagName.hs
@@ -55,6 +55,7 @@ instance Show FlagNameParseFail where
 newtype FlagName =
   FlagName ByteString
   deriving (Typeable,Data,Generic,Hashable,Binary,NFData)
+instance HasStructuralInfo FlagName
 instance Eq FlagName where
     x == y = (compare x y) == EQ
 instance Ord FlagName where