+++ src/Stack/Types/GhcPkgId.hs
@@ -52,8 +52,8 @@ instance Binary GhcPkgId where
         4 <- getWord8
         7 <- getWord8
         fmap GhcPkgId get
-instance NFData GhcPkgId where
-    rnf = genericRnf
+instance NFData GhcPkgId
+instance HasStructuralInfo GhcPkgId
 
 instance Show GhcPkgId where
   show = show . ghcPkgIdString