+++ src/Stack/Upgrade.hs
@@ -14,6 +14,7 @@ import qualified Data.Map                    as Map
 import           Data.Monoid                 ((<>))
 import qualified Data.Monoid
 import qualified Data.Set                    as Set
+import           Development.GitRev          (gitHash)
 import           Network.HTTP.Client.Conduit (HasHttpManager, getHttpManager)
 import           Path
 import qualified Paths_stack as Paths
@@ -27,6 +28,7 @@ import           Stack.Types
 import           Stack.Types.Internal
 import           Stack.Types.StackT
 import           System.IO.Temp              (withSystemTempDirectory)
+import           System.Process              (readProcess)
 import           System.Process.Run
 
 upgrade :: (MonadIO m, MonadMask m, MonadReader env m, HasConfig env, HasHttpManager env, MonadLogger m, HasTerminal env, HasReExec env, HasLogLevel env, MonadBaseControl IO m)
@@ -38,16 +40,21 @@ upgrade gitRepo mresolver = withSystemTempDirectory "stack-upgrade" $ \tmp' -> d
     tmp <- parseAbsDir tmp'
     mdir <- case gitRepo of
       Just repo -> do
-        $logInfo "Cloning stack"
-        runIn tmp "git" menv
-            [ "clone"
-            , repo
-            , "stack"
-            , "--depth"
-            , "1"
-            ]
-            Nothing
-        return $ Just $ tmp </> $(mkRelDir "stack")
+        remote <- liftIO $ readProcess "git" ["ls-remote", repo, "master"] []
+        let latestCommit = head . words $ remote
+        if (latestCommit == $gitHash) then do
+          $logInfo "Already up-to-date, no upgrade required"
+          return Nothing
+        else do $logInfo "Cloning stack"
+                runIn tmp "git" menv
+                    [ "clone"
+                    , repo
+                    , "stack"
+                    , "--depth"
+                    , "1"
+                    ]
+                    Nothing
+                return $ Just $ tmp </> $(mkRelDir "stack")
       Nothing -> do
         updateAllIndices menv
         caches <- getPackageCaches menv